requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
botocore==1.22.10
certifi==2021.5.30
charset-normalizer==2.0.6
cloudinary==1.26.0
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.5
Django==3.2.8
django-bootstrap-v5==1.0.6
django-bootstrap3==15.0.0
django-bootstrap4==3.0.1
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.0
django-heroku==0.3.1
django-registration==3.2
django-registration-redux==2.9
django-tinymce==3.3.0
djangorestframework==3.12.4
drf-yasg==1.20.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.2
importlib-metadata==3.10.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.21.3
openpyxl==3.0.9
packaging==21.2
pandas==1.3.4
Pillow==8.3.2
psycopg2==2.9.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.5
pyttsx3==2.90
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.1
scipy==1.7.1
six==1.16.0
soupsieve==2.2.1
spotipy==2.19.0
sqlparse==0.4.2
threadpoolctl==3.0.0
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
zipp==3.6.0
